refactor(macos): extract ChatView empty state and error banners into separate files
