+++ clients/macos/vellum-assistant/Features/Chat/ChatEmptyStateView.swift
@@ -1,0 +1,196 @@
+import SwiftUI
+import VellumAssistantShared
+
+/// Empty state shown when a ChatView thread has no messages.
+///
+/// Manages its own animation state (fade-in/scale) and randomises
+/// the greeting + placeholder text each time it appears. Embeds
+/// a `ComposerView` so the user can immediately start typing.
+struct ChatEmptyStateView: View {
+    @Binding var inputText: String
+    let hasAPIKey: Bool
+    let isSending: Bool
+    let isRecording: Bool
+    let suggestion: String?
+    let pendingAttachments: [ChatAttachment]
+    let onSend: () -> Void
+    let onStop: () -> Void
+    let onAcceptSuggestion: () -> Void
+    let onAttach: () -> Void
+    let onRemoveAttachment: (String) -> Void
+    let onPaste: () -> Void
+    let onMicrophoneToggle: () -> Void
+    @Binding var editorContentHeight: CGFloat
+    @Binding var isComposerExpanded: Bool
+
+    @State private var visible = false
+    @State private var title: String = titles.randomElement()!
+    @State private var placeholder: String = placeholderTexts.randomElement()!
+
+    // MARK: - Greeting Data
+
+    static let defaultGreetings = [
+        "What are we working on?",
+        "I'm here whenever you need me.",
+        "What's on your mind?",
+        "Let's make something happen.",
+        "Ready when you are.",
+    ]
+
+    static var titles: [String] {
+        let custom = IdentityInfo.loadGreetings()
+        return custom.isEmpty ? defaultGreetings : custom
+    }
+
+    static let placeholderTexts = [
+        "Ask me anything...",
+        "Tell me what you need...",
+        "Say the word...",
+        "Go ahead, I'm listening...",
+        "Type or hold Fn to talk...",
+    ]
+
+    // MARK: - Body
+
+    var body: some View {
+        VStack(spacing: 0) {
+            Spacer()
+            Spacer()
+
+            Text("🍃")
+                .font(.system(size: 48))
+                .opacity(visible ? 1 : 0)
+                .scaleEffect(visible ? 1 : 0.8)
+                .padding(.bottom, VSpacing.lg)
+
+            Text(title)
+                .font(.custom("Fraunces", size: 28).weight(.regular))
+                .foregroundColor(VColor.textSecondary)
+                .multilineTextAlignment(.center)
+                .frame(maxWidth: 500)
+                .opacity(visible ? 1 : 0)
+                .offset(y: visible ? 0 : 8)
+                .padding(.horizontal, VSpacing.xl)
+                .padding(.bottom, VSpacing.xl)
+
+            ComposerView(
+                inputText: $inputText,
+                hasAPIKey: hasAPIKey,
+                isSending: isSending,
+                isRecording: isRecording,
+                suggestion: suggestion,
+                pendingAttachments: pendingAttachments,
+                onSend: onSend,
+                onStop: onStop,
+                onAcceptSuggestion: onAcceptSuggestion,
+                onAttach: onAttach,
+                onRemoveAttachment: onRemoveAttachment,
+                onPaste: onPaste,
+                onMicrophoneToggle: onMicrophoneToggle,
+                placeholderText: placeholder,
+                editorContentHeight: $editorContentHeight,
+                isComposerExpanded: $isComposerExpanded
+            )
+            .frame(maxWidth: 500)
+            .opacity(visible ? 1 : 0)
+            .offset(y: visible ? 0 : 10)
+
+            Spacer()
+            Spacer()
+            Spacer()
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .onAppear {
+            withAnimation(.easeOut(duration: 0.5)) {
+                visible = true
+            }
+        }
+        .onDisappear {
+            visible = false
+        }
+    }
+
+}
+
+// MARK: - Temporary Chat Empty State
+
+/// Variant shown for temporary (non-persistent) chats.
+struct ChatTemporaryChatEmptyStateView: View {
+    @Binding var inputText: String
+    let hasAPIKey: Bool
+    let isSending: Bool
+    let isRecording: Bool
+    let suggestion: String?
+    let pendingAttachments: [ChatAttachment]
+    let onSend: () -> Void
+    let onStop: () -> Void
+    let onAcceptSuggestion: () -> Void
+    let onAttach: () -> Void
+    let onRemoveAttachment: (String) -> Void
+    let onPaste: () -> Void
+    let onMicrophoneToggle: () -> Void
+    @Binding var editorContentHeight: CGFloat
+    @Binding var isComposerExpanded: Bool
+
+    var body: some View {
+        VStack(spacing: 0) {
+            Spacer()
+            Spacer()
+
+            Text("🍃")
+                .font(.system(size: 48))
+                .padding(.bottom, VSpacing.lg)
+
+            Text("Temporary Chat")
+                .font(.system(size: 28, weight: .medium))
+                .foregroundColor(VColor.textSecondary)
+                .multilineTextAlignment(.center)
+                .padding(.bottom, VSpacing.sm)
+
+            Text("Memory is disabled for this chat, and it won\u{2019}t appear in your history.")
+                .font(VFont.body)
+                .foregroundColor(VColor.textMuted)
+                .multilineTextAlignment(.center)
+                .frame(maxWidth: 400)
+                .padding(.horizontal, VSpacing.xl)
+                .padding(.bottom, VSpacing.xxl)
+
+            ComposerView(
+                inputText: $inputText,
+                hasAPIKey: hasAPIKey,
+                isSending: isSending,
+                isRecording: isRecording,
+                suggestion: suggestion,
+                pendingAttachments: pendingAttachments,
+                onSend: onSend,
+                onStop: onStop,
+                onAcceptSuggestion: onAcceptSuggestion,
+                onAttach: onAttach,
+                onRemoveAttachment: onRemoveAttachment,
+                onPaste: onPaste,
+                onMicrophoneToggle: onMicrophoneToggle,
+                placeholderText: "Ask anything...",
+                editorContentHeight: $editorContentHeight,
+                isComposerExpanded: $isComposerExpanded
+            )
+
+            Spacer()
+            Spacer()
+            Spacer()
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .background(
+            RadialGradient(
+                gradient: Gradient(colors: [
+                    VColor.accent.opacity(0.07),
+                    VColor.accent.opacity(0.02),
+                    Color.clear,
+                ]),
+                center: .center,
+                startRadius: 20,
+                endRadius: 350
+            )
+            .offset(y: -40)
+        )
+    }
+}
